requirements.txt
streamlit
openai
boto3
python-terraform
python-dotenv
